Refactor and improve build output (#2202)
